+++ packer/rpc/post_processor.go
@@ -19,6 +19,7 @@ type PostProcessorServer struct {
 
 type PostProcessorProcessResponse struct {
 	Err        error
+	Keep       bool
 	RPCAddress string
 }
 
@@ -33,30 +34,30 @@ func (p *postProcessor) Configure(raw interface{}) (err error) {
 	return
 }
 
-func (p *postProcessor) PostProcess(ui packer.Ui, a packer.Artifact) (packer.Artifact, error) {
+func (p *postProcessor) PostProcess(ui packer.Ui, a packer.Artifact) (packer.Artifact, bool, error) {
 	server := rpc.NewServer()
 	RegisterArtifact(server, a)
 	RegisterUi(server, ui)
 
 	var response PostProcessorProcessResponse
 	if err := p.client.Call("PostProcessor.PostProcess", serveSingleConn(server), &response); err != nil {
-		return nil, err
+		return nil, false, err
 	}
 
 	if response.Err != nil {
-		return nil, response.Err
+		return nil, false, response.Err
 	}
 
 	if response.RPCAddress == "" {
-		return nil, nil
+		return nil, false, nil
 	}
 
 	client, err := rpc.Dial("tcp", response.RPCAddress)
 	if err != nil {
-		return nil, err
+		return nil, false, err
 	}
 
-	return Artifact(client), nil
+	return Artifact(client), response.Keep, nil
 }
 
 func (p *PostProcessorServer) Configure(raw *interface{}, reply *error) error {
@@ -76,7 +77,7 @@ func (p *PostProcessorServer) PostProcess(address string, reply *PostProcessorPr
 
 	responseAddress := ""
 
-	artifact, err := p.p.PostProcess(&Ui{client}, Artifact(client))
+	artifact, keep, err := p.p.PostProcess(&Ui{client}, Artifact(client))
 	if err == nil && artifact != nil {
 		server := rpc.NewServer()
 		RegisterArtifact(server, artifact)
@@ -89,6 +90,7 @@ func (p *PostProcessorServer) PostProcess(address string, reply *PostProcessorPr
 
 	*reply = PostProcessorProcessResponse{
 		Err:        err,
+		Keep:       keep,
 		RPCAddress: responseAddress,
 	}